+++ src/metrics/family.rs
@@ -97,6 +97,7 @@ use std::sync::{Arc, RwLock, RwLockReadGuard};
 /// # assert_eq!(expected, String::from_utf8(buffer).unwrap());
 /// ```
 // TODO: Consider exposing hash algorithm.
+#[derive(Debug)]
 pub struct Family<S, M, C = fn() -> M> {
     metrics: Arc<RwLock<HashMap<S, M>>>,
     /// Function that when called constructs a new metric.
@@ -144,9 +145,9 @@ pub trait MetricConstructor<M> {
 /// ```
 /// # use prometheus_client::metrics::family::{Family};
 /// # use prometheus_client::metrics::histogram::Histogram;
-/// let custom_buckets = vec![0.0, 10.0, 100.0];
+/// let custom_buckets = [0.0, 10.0, 100.0];
 /// let metric = Family::<(), Histogram, _>::new_with_constructor(|| {
-///     Histogram::new(custom_buckets.clone().into_iter())
+///     Histogram::new(custom_buckets.into_iter())
 /// });
 /// # metric.get_or_create(&());
 /// ```