+++ src/metrics/histogram.rs
@@ -14,8 +14,24 @@ use std::sync::{Arc, Mutex, MutexGuard};
 /// let histogram = Histogram::new(exponential_buckets(1.0, 2.0, 10));
 /// histogram.observe(4.2);
 /// ```
+///
+/// [`Histogram`] does not implement [`Default`], given that the choice of
+/// bucket values depends on the situation [`Histogram`] is used in. As an
+/// example, to measure HTTP request latency, the values suggested in the
+/// Golang implementation might work for you:
+///
+/// ```
+/// # use prometheus_client::metrics::histogram::Histogram;
+/// // Default values from go client(https://github.com/prometheus/client_golang/blob/5d584e2717ef525673736d72cd1d12e304f243d7/prometheus/histogram.go#L68)
+/// let custom_buckets = [
+///    0.005, 0.01, 0.025, 0.05, 0.1, 0.25, 0.5, 1.0, 2.5, 5.0, 10.0,
+/// ];
+/// let histogram = Histogram::new(custom_buckets.into_iter());
+/// histogram.observe(4.2);
+/// ```
 // TODO: Consider using atomics. See
 // https://github.com/tikv/rust-prometheus/pull/314.
+#[derive(Debug)]
 pub struct Histogram {
     inner: Arc<Mutex<Inner>>,
 }
@@ -28,6 +44,7 @@ impl Clone for Histogram {
     }
 }
 
+#[derive(Debug)]
 pub(crate) struct Inner {
     // TODO: Consider allowing integer observe values.
     sum: f64,